Force-app: renamed app packages.
