+++ src/main/java/br/com/schumaker/force/app/model/UseInjectedBeans.java
@@ -1,10 +1,10 @@
-package br.com.schumaker.octopus.app.model;
+package br.com.schumaker.force.app.model;
 
-import br.com.schumaker.octopus.app.view.ProductDTO;
-import br.com.schumaker.octopus.framework.annotations.bean.Component;
-import br.com.schumaker.octopus.framework.annotations.bean.Inject;
-import br.com.schumaker.octopus.framework.model.Mapper;
-import br.com.schumaker.octopus.framework.web.http.HttpRestTemplate;
+import br.com.schumaker.force.app.view.ProductDTO;
+import br.com.schumaker.force.framework.annotations.bean.Component;
+import br.com.schumaker.force.framework.annotations.bean.Inject;
+import br.com.schumaker.force.framework.model.Mapper;
+import br.com.schumaker.force.framework.web.http.HttpRestTemplate;
 
 /**
  * The UseInjectedBeans class.
