+++ src/main/java/br/com/schumaker/force/app/view/ProductView.java
@@ -1,4 +1,4 @@
-package br.com.schumaker.octopus.app.view;
+package br.com.schumaker.force.app.view;
 
 import java.math.BigInteger;
